backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
bcrypt
beautifulsoup4
black
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
click
cryptography
distro
dnspython
ecdsa
email-validator
emergentintegrations
et_xmlfile
fastapi
fastuuid
filelock
flake8
frozenlist
fsspec
gdown
git-filter-repo
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
google_search_results
googleapis-common-protos
grpcio
grpcio-status
h11
hf-xet
httpcore
httplib2
httpx
huggingface_hub
idna
importlib_metadata
iniconfig
isort
Jinja2
jiter
jmespath
jq
jsonschema
jsonschema-specifications
librt
litellm
lxml
markdown-it-py
MarkupSafe
mccabe
mdurl
motor
multidict
mypy
mypy_extensions
numpy
oauthlib
openai
openpyxl
packaging
pandas
passlib
pathspec
pdf2image
pdfminer.six
pdfplumber
pillow
platformdirs
pluggy
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic_core
pyflakes
Pygments
PyJWT
pymongo
PyMuPDF
pyparsing
pypdfium2
PySocks
pytesseract
pytest
python-dateutil
python-docx
python-dotenv
python-jose
python-multipart
python-pptx
python-slugify
pytokens
pytz
PyYAML
RapidFuzz
referencing
regex
reportlab
requests
requests-oauthlib
rich
rpds-py
rsa
s3transfer
s5cmd
shellingham
six
sniffio
soupsieve
starlette
stripe
tenacity
text-unidecode
tiktoken
tokenizers
tqdm
typer
typer-slim
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
watchfiles
websockets
xlsxwriter
yarl
zipp
